// frontend/src/groups/serviceaccount/__generated__/ServiceAccountNamespaceMembershipsFragment.graphql.ts
export type ServiceAccountNamespaceMembershipsFragment$data = {
  readonly id: string;
  readonly namespaceMemberships: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly metadata: {
          readonly updatedAt: any;
        };
        readonly namespace: {
          readonly fullPath: string;
        } | null | undefined;
        readonly role: {
          readonly name: string;
        };
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "ServiceAccountNamespaceMembershipsFragment";
};
export type ServiceAccountNamespaceMembershipsFragment$key = {
  readonly " $data"?: ServiceAccountNamespaceMembershipsFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"ServiceAccountNamespaceMembershipsFragment">;
};